+++ src/main/kotlin/com/grepp/nbe1_3_team9/domain/entity/finance/Expense.kt
@@ -1,4 +1,4 @@
-package com.grepp.nbe1_3_team9.domain.entity
+package com.grepp.nbe1_3_team9.domain.entity.finance
 
 import com.grepp.nbe1_3_team9.domain.entity.group.Group
 import jakarta.persistence.*
