+++ src/main/kotlin/com/grepp/nbe1_3_team9/domain/repository/finance/AccountBookRepository.kt
@@ -1,6 +1,6 @@
 package com.grepp.nbe1_3_team9.domain.repository.finance
 
-import com.grepp.nbe1_3_team9.domain.entity.Expense
+import com.grepp.nbe1_3_team9.domain.entity.finance.Expense
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface AccountBookRepository : JpaRepository<Expense, Long> {
